Used cascaded value system to index item sets.
