+++ src/ValueExtractor/ItemValueExtractor.php
@@ -2,6 +2,7 @@
 
 /*
  * Copyright BibLibre, 2016-2017
+ * Copyright Daniel Berthereau, 2018
  *
  * This software is governed by the CeCILL license under French law and abiding
  * by the rules of distribution of free software.  You can use, modify and/ or
@@ -32,11 +33,11 @@ namespace Solr\ValueExtractor;
 use Omeka\Api\Manager as ApiManager;
 use Omeka\Api\Representation\ItemRepresentation;
 use Omeka\Api\Representation\AbstractResourceRepresentation;
-use Omeka\Api\Representation\AbstractResourceEntityRepresentation;
-use Omeka\Api\Representation\ValueRepresentation;
 
 class ItemValueExtractor implements ValueExtractorInterface
 {
+    use ValueExtractorTrait;
+
     /**
      * @var ApiManager
      */
@@ -188,60 +189,5 @@ class ItemValueExtractor implements ValueExtractorInterface
         }
 
         return $extractedValue;
-    }
-
-    /**
-     * Extract the values of the given property of the given item.
-     * If a value is a resource, then this method is called recursively with
-     * the source part after the slash as $source.
-     *
-     * @param AbstractResourceEntityRepresentation $representation Item
-     * @param string $source Property (RDF term).
-     * @return string[] Human-readable values.
-     */
-    protected function extractPropertyValue(AbstractResourceEntityRepresentation $representation, $source)
-    {
-        @list($property, $subProperty) = explode('/', $source, 2);
-        $extractedValue = [];
-        /* @var $values ValueRepresentation[] */
-        $values = $representation->value($property, ['all' => true, 'default' => []]);
-        foreach ($values as $value) {
-            $type = $value->type();
-            if ($type === 'literal' || $type == 'uri') {
-                $extractedValue[] = (string) $value;
-            } elseif ('resource' === explode(':', $type)[0]) {
-                $this->extractPropertyResourceValue($extractedValue, $value, $subProperty);
-            }
-        }
-
-        return $extractedValue;
-    }
-
-    /**
-     * Extracts value(s) from resource-type value and adds them to already
-     * extracted values (passed by reference).
-     *
-     * @param array $extractedValues Already extracted values.
-     * @param ValueRepresentation $value Resource-type value from which to
-     * extract searched values.
-     * @param string|null $property RDF term representing the property to
-     * extract. If null, get the displayTitle() value.
-     */
-    protected function extractPropertyResourceValue(
-        array &$extractedValues,
-        ValueRepresentation $value,
-        $property
-    ) {
-        if (isset($property)) {
-            $extractedValues = array_merge(
-                $extractedValues,
-                $this->extractPropertyValue($value->valueResource(), $property)
-            );
-        } else {
-            $resourceTitle = $value->valueResource()->displayTitle('');
-            if (!empty($resourceTitle)) {
-                $extractedValues[] = $resourceTitle;
-            }
-        }
     }
 }
